Implementing signup, login, and role-based authentication in the project
