--- src/auth/guard/jwt-auth.guard.js
@@ -1,8 +1,0 @@
-import { AuthGuard } from '@nestjs/passport';
-// import { JwtGuard } from '../auth/guard';
-
-export class JwtAuthGuard extends AuthGuard('jwt') {
-  constructor() {
-    super();
-  }
-}
