--- src/auth/guard/jwt.auth.guard.ts
@@ -1,30 +1,0 @@
-import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { JwtService } from '@nestjs/jwt';
-import { AuthService } from '../auth.service';
-
-@Injectable()
-export class JwtAuthGuard implements CanActivate {
-  constructor(
-    private jwtService: JwtService,
-    private authService: AuthService,
-  ) {}
-
-  async canActivate(context: ExecutionContext): Promise<boolean> {
-    const request = context.switchToHttp().getRequest();
-    const token = request.headers.authorization?.split(' ')[1];
-
-    if (!token) {
-      return false;
-    }
-
-    const decoded = this.jwtService.decode(token);
-    const user = await this.authService.validateUser(decoded);
-
-    if (!user) {
-      return false;
-    }
-
-    request.user = user;
-    return true;
-  }
-}
