+++ src/auth/guard/role.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { ROLES_KEY } from 'src/decorator/roles.decorator';
-import { Role } from '../roles.enum';
+import { Role } from 'src/enum/role.enum';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
