--- src/auth/roles.enum.ts
@@ -1,9 +1,0 @@
-export enum Role {
-  SUPER_ADMIN = 1,
-  SUB_ADMIN = 2,
-  BRANCH_MANAGER = 3,
-  DELIVERY_BOY = 4,
-  CUSTOMER = 5,
-  WORKSHOP_MANAGER = 6,
-  VENDOR_RESELLER = 7,
-}
