+++ src/auth/strategy/jwt-strategy.ts
@@ -13,7 +13,6 @@ export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
   }
 
   async validate(payload: any) {
-    console.log("validate", payload);
     return payload;
   }
 }
